+++ app/Http/Controllers/TerrainController.php
@@ -9,18 +9,21 @@ use App\Http\Resources\TerrainResource;
 use App\Models\Terrain;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Resources\Json\JsonResource;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
 
 class TerrainController extends Controller
 {
     public function index(): JsonResource
     {
-        $terrain = Terrain::with('type', 'proprietaire')->get();
+        $query = Terrain::with('type', 'proprietaire');
+        $terrain = Auth::user()->can('viewAny', Terrain::class) ? $query->get() : $query->owner()->get();
         return TerrainListResource::collection($terrain);
     }
 
     public function store(StoreRequest $request): JsonResponse
     {
+        $this->authorize('create', Terrain::class);
         $request->validated();
         $terrain = Terrain::make($request->all());
         $terrain->genererCode();
@@ -30,11 +33,13 @@ class TerrainController extends Controller
 
     public function show(Terrain $terrain): JsonResource
     {
+        $this->authorize('view', $terrain);
         return TerrainResource::make($terrain);
     }
 
     public function update(UpdateRequest $updateRequest, Terrain $terrain): JsonResponse
     {
+        $this->authorize('update', $terrain);
         $updateRequest->validated();
         $terrain->update($updateRequest->all());
         return response()->json("Le terrain a été modifié avec succès.");
@@ -42,6 +47,7 @@ class TerrainController extends Controller
 
     public function destroy(Terrain $terrain)
     {
+        $this->authorize('delete', $terrain);
         $terrain->delete();
         return response()->json("Le terrain $terrain->nom a été supprimé avec succès.");
     }
